Ensure `.json` and `.xml` reports go to report dir
The XML and JSON report files that were generated were _always_ being
written to the root of the current directory instead of the report
directory that the application _told_ the user that it was writing to.

i.e:

What it told the user: `<current-dir>/qark/report/Report.json`
What it actually did: `<current-dir/Report.json`

This fixes that.